Jenkins _ Test Edge

+++ src/main/java/commons/BaseTest.java
@@ -152,7 +152,7 @@ public class BaseTest {
             String pathFolderDownload = GlobalConstants.PROJECT_PATH + File.separator + folderName;
             File file = new File(pathFolderDownload);
             File[] listOfFiles = file.listFiles();
-            if (listOfFiles.length != 0) {
+            if (listOfFiles != null && listOfFiles.length != 0) {
                 for (int i = 0; i < listOfFiles.length; i++) {
                     if (listOfFiles[i].isFile() && !listOfFiles[i].getName().equals("environment.properties")) {
                         new File(listOfFiles[i].toString()).delete();
